DMVM-142(feature): feat onboarding reducer
